refactor(etl/database): изменить типы данных для аналитической PostgreSQL
• Заменить SERIAL на BIGSERIAL для первичных ключей users и activities
• Изменить INTEGER на BIGINT для внешних ключей user_id и activity_id
• Упростить объявление полей в конфигурации Pydantic, убрав default=
• Добавить переменную окружения ETL_INTERVAL_MINUTES в workflow развертывания

+++ etl_service/ddl.py
@@ -4,14 +4,14 @@
 # SQL команды для создания таблиц
 CREATE_TABLE_USERS = """
     CREATE TABLE IF NOT EXISTS users (
-        id SERIAL PRIMARY KEY,
+        id BIGSERIAL PRIMARY KEY,
         nickname VARCHAR(255)
     )
 """
 
 CREATE_TABLE_ACTIVITIES = """
     CREATE TABLE IF NOT EXISTS activities (
-        id SERIAL PRIMARY KEY,
+        id BIGSERIAL PRIMARY KEY,
         name VARCHAR(255) NOT NULL,
         unit VARCHAR(50) NOT NULL,
         calories_per_unit DECIMAL(5,3)
@@ -21,7 +21,7 @@ CREATE_TABLE_ACTIVITIES = """
 CREATE_TABLE_WEIGHT_DATA = """
     CREATE TABLE IF NOT EXISTS weight_data (
         id SERIAL PRIMARY KEY,
-        user_id INTEGER NOT NULL,
+        user_id BIGINT NOT NULL,
         weight DECIMAL(3,1) NOT NULL,
         date DATE NOT NULL,
         FOREIGN KEY (user_id) REFERENCES users (id)
@@ -31,8 +31,8 @@ CREATE_TABLE_WEIGHT_DATA = """
 CREATE_TABLE_ACTIVITY_DATA = """
     CREATE TABLE IF NOT EXISTS activity_data (
         id SERIAL PRIMARY KEY,
-        user_id INTEGER NOT NULL,
-        activity_id INTEGER NOT NULL,
+        user_id BIGINT NOT NULL,
+        activity_id BIGINT NOT NULL,
         date DATE NOT NULL,
         value DECIMAL(6,2) NOT NULL,
         calories INTEGER NOT NULL,
